// src/services/database.types.ts
export type Database = {
  public: {
    Tables: {
      Issue: {
        Row: {
          category: string
          description: string
          id: string
          image: string | null
          location: string | null
          priority: Database["public"]["Enums"]["Priority"]
          reportedAt: string
          reportedBy: string | null
          resolutionNotes: string | null
          resolvedAt: string | null
          resolvedBy: string | null
          status: Database["public"]["Enums"]["Status"]
          synced: boolean
          title: string
        }
        Insert: {
          category: string
          description: string
          id: string
          image?: string | null
          location?: string | null
          priority?: Database["public"]["Enums"]["Priority"]
          reportedAt?: string
          reportedBy?: string | null
          resolutionNotes?: string | null
          resolvedAt?: string | null
          resolvedBy?: string | null
          status?: Database["public"]["Enums"]["Status"]
          synced?: boolean
          title: string
        }
        Update: {
          category?: string
          description?: string
          id?: string
          image?: string | null
          location?: string | null
          priority?: Database["public"]["Enums"]["Priority"]
          reportedAt?: string
          reportedBy?: string | null
          resolutionNotes?: string | null
          resolvedAt?: string | null
          resolvedBy?: string | null
          status?: Database["public"]["Enums"]["Status"]
          synced?: boolean
          title?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      Priority: "urgent" | "high" | "medium" | "low"
      Status: "pending" | "in_progress" | "resolved"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type PublicSchema = Database[Extract<keyof Database, "public">]

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
        PublicSchema["Views"])
    ? (PublicSchema["Tables"] &
        PublicSchema["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R
      }
      ? R
      : never
    : never
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
    ? PublicSchema["Enums"][PublicEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof PublicSchema["CompositeTypes"]
    | { schema: keyof Database },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends { schema: keyof Database }
  ? Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof PublicSchema["CompositeTypes"]
    ? PublicSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never